Changed Decimal to Integer

+++ AB2020/src/de/die_gfi/philipp/Schachbrett.java
@@ -1,6 +1,7 @@
 package de.die_gfi.philipp;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.text.DecimalFormat;
 import java.text.DecimalFormatSymbols;
 import java.util.Locale;
@@ -10,7 +11,7 @@ public class Schachbrett {
     static DecimalFormat format = new DecimalFormat("#.###E0", DecimalFormatSymbols.getInstance(Locale.ROOT));
 
     public static void main(String[] args) {
-        BigDecimal bigNum = BigDecimal.ONE;
+        BigInteger bigNum = BigInteger.ONE;
         for (int count = 0; count < 64; count++) {
             if (count % 8 == 0) {
                 if (count != 0) {
@@ -19,7 +20,7 @@ public class Schachbrett {
                 System.out.print("\n" + "+--------".repeat(8) + "+\n" + "|        ".repeat(8) + "|\n|");
             }
             System.out.printf("%8s|", format.format(bigNum));
-            bigNum = bigNum.multiply(BigDecimal.valueOf(2));
+            bigNum = bigNum.multiply(BigInteger.valueOf(2));
         }
         System.out.println("\n" + "|        ".repeat(8) + "|\n" + "+--------".repeat(8) + "+");
     }
